+++ activity/jd_global.js
